style(pipeline): clear ruff 0.16 findings; ignore BLE001; pin ruff
CI lint had been red since July: ruff is unpinned, so runners pulled a newer
ruff (0.16.5) whose broader default rules flagged 48 issues in unchanged code
(first push since July surfaced it). Tests and the web build were fine.

- Auto-fixes: import sorting (I001), datetime.UTC (UP017), range() starts
  (PIE808), and other safe rewrites.
- By hand: bind the Open-Meteo loop var in the closure (B023), drop the now
  redundant fromisoformat "Z" replace on py311+ (FURB162), explicit UTC date
  for the recent-tier cutoff (DTZ011), drop an int() over math.floor (RUF046),
  parenthesize a multi-line message (ISC004), collapse a nested if (SIM102).
- noqa with rationale where a naive parse is intentional (DTZ007: IST/local
  time attached right after) and one defensive file-skip (S112).
- Ignore BLE001 globally: the ingest/build broad excepts are deliberate
  resilience (one flaky sensor must not abort a refresh).
- Pin ruff to >=0.16.5,<0.17 so new-rule minor bumps can't silently re-red CI.

No behavior change; 129 tests pass, ruff clean.

+++ pipeline/aqi/compute.py
@@ -44,7 +44,7 @@ def _truncate(value: float, decimals: int) -> float:
 
 
 def _round_half_up(value: float) -> int:
-    return int(math.floor(value + 0.5))
+    return math.floor(value + 0.5)
 
 
 def ugm3_to_ppb(pollutant: str, value_ugm3: float) -> float:
